Maintenance, Tweaks #FIX: Fixed serialization problems after renaming by introducing a new compatibility error handler. #CHANGE: RenderTarget.Bind is no longer being called as part of the rendering process. The graphics backend API should be bind-less and definitely NOT state-machine like.
