+++ Duality/Backend/DefaultOpenTK/DefaultOpenTKBackend.cs
@@ -32,6 +32,9 @@ namespace Duality.Backend.DefaultOpenTK
 			this.currentDevice = device;
 			this.renderStats = stats;
 
+			// Prepare the target surface for rendering
+			RenderTarget.Bind(options.Target);
+
 			if (this.primaryVBO == 0) GL.GenBuffers(1, out this.primaryVBO);
 			GL.BindBuffer(BufferTarget.ArrayBuffer, this.primaryVBO);
