+++ DualityEditorPlugins/ProjectView/TreeModels/ModelItems.cs
@@ -187,7 +187,7 @@ namespace Duality.Editor.Plugins.ProjectView.TreeModels
 		{
 			string[] fileNameSplit = Path.GetFileNameWithoutExtension(path).Split('.');
 
-			this.res.Path = path;
+			this.res = new ContentRef<Resource>(null, path);
 			this.resType = Resource.GetTypeByFileName(path);
 			this.ApplyPathToName();
 		}
