+++ DualityPlugins/Compatibility/Compatibility.csproj
@@ -43,7 +43,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="CorePlugin.cs" />
-    <Compile Include="MovedTextureSizeModeErrorHandler.cs" />
+    <Compile Include="MovedTypesV2ErrorHandler.cs" />
     <Compile Include="RigidBodyJointErrorHandler.cs" />
     <Compile Include="RigidBodyRendererErrorHandler.cs" />
     <Compile Include="OldNamespacesErrorHandler.cs" />
