+++ DualityPlugins/Compatibility/MovedTypesV2ErrorHandler.cs
@@ -1,0 +1,33 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+
+using Duality;
+using Duality.Serialization;
+
+namespace Duality.Plugins.Compatibility
+{
+	public class MovedTypesV2ErrorHandler : SerializeErrorHandler
+	{
+		public override void HandleError(SerializeError error)
+		{
+			ResolveTypeError resolveTypeError = error as ResolveTypeError;
+			if (resolveTypeError != null)
+			{
+				string fixedTypeId = resolveTypeError.TypeId.Replace('+', '.');
+
+				if (fixedTypeId.EndsWith("Duality.Resources.Texture.SizeMode"))
+					resolveTypeError.ResolvedType = typeof(Duality.Drawing.TextureSizeMode);
+				else if (fixedTypeId.EndsWith("Duality.Resources.BatchInfo"))
+					resolveTypeError.ResolvedType = typeof(Duality.Drawing.BatchInfo);
+				else if (fixedTypeId.EndsWith("Duality.Resources.BatchInfo.DirtyFlag"))
+					resolveTypeError.ResolvedType = typeof(Duality.Drawing.BatchInfo).GetNestedType("DirtyFlag", ReflectionHelper.BindAll);
+				else if (fixedTypeId.EndsWith("Duality.Drawing.DefaultRendererVisibilityStrategy"))
+					resolveTypeError.ResolvedType = typeof(Duality.Components.DefaultRendererVisibilityStrategy);
+			}
+
+			return;
+		}
+	}
+}
